+++ src/games/space-invaders/src/interfaces/bullet.interface.ts
@@ -1,0 +1,8 @@
+export interface IBulletConstructor {
+  scene: Phaser.Scene;
+  bulletProperties: { speed: number };
+  x: number;
+  y: number;
+  texture: string;
+  frame?: string | number;
+}