+++ src/games/space-invaders/src/scenes/boot-scene.ts
@@ -1,10 +1,3 @@
-/**
- * @author       Digitsensitive <digit.sensitivee@gmail.com>
- * @copyright    2018 - 2019 digitsensitive
- * @description  Space Invaders: Boot Scene
- * @license      Digitsensitive
- */
-
 import { AnimationHelper } from '../helpers/animation-helper';
 
 export class BootScene extends Phaser.Scene {
@@ -26,7 +19,7 @@ export class BootScene extends Phaser.Scene {
     // pass value to change the loading bar fill
     this.load.on(
       'progress',
-      function (value) {
+      function (value: number) {
         this.progressBar.clear();
         this.progressBar.fillStyle(0xfff6d3, 1);
         this.progressBar.fillRect(
@@ -54,11 +47,7 @@ export class BootScene extends Phaser.Scene {
     );
 
     // load out package
-    this.load.pack(
-      'preload',
-      './src/games/space-invaders/assets/pack.json',
-      'preload'
-    );
+    this.load.pack('preload', './assets/pack.json', 'preload');
   }
 
   update(): void {