+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Refactor `coin-runner`
 - Refactor `flappy-bird`
 - Refactor `snake`
+- Refactor `space-invaders`
 
 ---